fix: divide go back button to reuse
- 뒤로 가기 버튼을 별도로 분리함으로써 재사용 가능하도록 변경
