+++ web-apps/client/src/components/tool-bar.js
@@ -1,35 +1,25 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useHistory } from 'react-router-dom';
-
-import ArrowBackIcon from '@material-ui/icons/ArrowBack';
-import { IconButton } from '@material-ui/core';
 
 import Actionbar from './action-bar';
+import GoBackButton from './goBackButton';
 
-const ActivityLayorToolbar = ({ title, buttons = [] }) => {
-  const history = useHistory();
-  const goBack = () => {
-    history.goBack();
-  };
-  const BackButtons = (
-    <IconButton aria-label='back' onClick={goBack}>
-      <ArrowBackIcon />
-    </IconButton>
-  );
+const ActivityLayorToolbar = ({title, buttons = []}) => {
   return (
-    <Actionbar leftArea={BackButtons} title={title} buttons={buttons} />
-  )
+    <Actionbar leftArea={<GoBackButton />} title={title} buttons={buttons} />
+  );
 };
 
 ActivityLayorToolbar.propTypes = {
   title: PropTypes.node.isRequired,
-  buttons: PropTypes.arrayOf(PropTypes.shape({
-    name: PropTypes.string,
-    href: PropTypes.string,
-    icon: PropTypes.node,
-    onClick: PropTypes.func,
-  })),
+  buttons: PropTypes.arrayOf(
+    PropTypes.shape({
+      name: PropTypes.string,
+      href: PropTypes.string,
+      icon: PropTypes.node,
+      onClick: PropTypes.func,
+    }),
+  ),
 };
 
 ActivityLayorToolbar.defaultProps = {
